requirements.txt
requests
zendriver
discord.py
markdownify